refactor: remove settings context
