+++ src/features/auth/context/useSettings.ts
@@ -1,0 +1,20 @@
+import { useQuery } from 'react-query'
+
+import { api } from 'api/api'
+import { SettingsResponse } from 'api/gen'
+import { QueryKeys } from 'libs/queryKeys'
+
+// arbitrary. Should not change that often though
+const STALE_TIME_APP_SETTINGS = 5 * 60 * 1000
+
+export const useSettings = () => {
+  const { data, isLoading } = useQuery<SettingsResponse>(
+    [QueryKeys.SETTINGS],
+    () => api.getNativeV1Settings(),
+    {
+      staleTime: STALE_TIME_APP_SETTINGS,
+    }
+  )
+
+  return { data, isLoading }
+}
